JP-3311 confirmation message add

+++ comu/com.energyict.comserver.comserver.core.extjs/src/main/web/js/mdc/app/controller/setup/LoadProfileTypes.js
@@ -47,7 +47,7 @@ Ext.define('Mdc.controller.setup.LoadProfileTypes', {
                 afterrender: this.loadStore
             },
             'loadProfileTypeSetup loadProfileTypeGrid': {
-                itemclick: this.loadGridItemDetail
+                select: this.loadGridItemDetail
             },
             'loadProfileTypeAddMeasurementTypesView': {
                 afterrender: this.measurementTypesLoad
@@ -128,50 +128,42 @@ Ext.define('Mdc.controller.setup.LoadProfileTypes', {
     },
 
     showConfirmationPanel: function () {
-        var grid = this.getLoadTypeGrid(),
+        var me = this,
+            grid = me.getLoadTypeGrid(),
             lastSelected = grid.getView().getSelectionModel().getLastSelected();
-        Ext.widget('messagebox', {
-            buttons: [
-                {
-                    xtype: 'button',
-                    text: 'Remove',
-                    action: 'removeloadprofiletypeconfirm',
-                    name: 'delete',
-                    ui: 'remove'
-                },
-                {
-                    xtype: 'button',
-                    text: 'Cancel',
-                    handler: function (button, event) {
-                        this.up('messagebox').hide();
-                    },
-                    ui: 'link'
-                }
-            ]
-        }).show({
-            title: 'Remove ' + lastSelected.getData().name + '?',
-            msg: 'This load profile type will no longer be available',
-            icon: Ext.MessageBox.WARNING
-        })
+
+        Ext.create('Uni.view.window.Confirmation').show({
+            msg: Uni.I18n.translate('deviceType.confirmWindow.removeMsg', 'MDC', 'This load profile type will no longer be available'),
+            title: Uni.I18n.translate('general.remove', 'MDC', 'Remove') + ' ' + lastSelected.get('name') + '?',
+            config: {
+                me: me
+            },
+            fn: me.confirmationPanelHandler
+        });
     },
 
-    deleteRecord: function (btn) {
-        var grid = this.getLoadTypeGrid(),
-            lastSelected = grid.getView().getSelectionModel().getLastSelected(),
-            me = this;
-        btn.up('messagebox').hide();
-        Ext.Ajax.request({
-            url: '/api/mds/loadprofiles/' + lastSelected.getData().id,
-            method: 'DELETE',
-            waitMsg: 'Removing...',
-            success: function () {
-                me.handleSuccessRequest('Load profile type was removed successfully');
-                me.store.load();
-            },
-            failure: function (result, request) {
-                me.handleFailureRequest(result, "Error during removing of load profile", 'retryremoveloadprofiletype');
-            }
-        });
+    confirmationPanelHandler: function (state, text, conf) {
+        var me = conf.config.me,
+            grid = me.getLoadTypeGrid(),
+            lastSelected = grid.getView().getSelectionModel().getLastSelected();
+
+        if (state === 'confirm') {
+            this.close();
+            Ext.Ajax.request({
+                url: '/api/mds/loadprofiles/' + lastSelected.getData().id,
+                method: 'DELETE',
+                waitMsg: 'Removing...',
+                success: function () {
+                    me.handleSuccessRequest('Load profile type was removed successfully');
+                    me.store.load();
+                },
+                failure: function (result, request) {
+                    me.handleFailureRequest(result, "Error during removing of load profile", 'retryremoveloadprofiletype');
+                }
+            });
+        } else if (state === 'cancel') {
+            this.close();
+        }
     },
 
     retrySubmit: function (btn) {
@@ -279,33 +271,33 @@ Ext.define('Mdc.controller.setup.LoadProfileTypes', {
         }).show();
     },
 
-/*    handleFailureRequest: function (response, headerText, retryAction) {
-        Ext.EventManager.stopPropagation('requestexception');
-        var errormsgs = response.responseText;
-        Ext.widget('messagebox', {
-            buttons: [
-                {
-                    text: 'Retry',
-                    action: retryAction,
-                    ui: 'remove'
-                },
-                {
-                    text: 'Cancel',
-                    action: 'cancel',
-                    ui: 'link',
-                    href: '#/administration/loadprofiletypes/',
-                    handler: function (button, event) {
-                        this.up('messagebox').hide();
-                    }
-                }
-            ]
-        }).show({
-            ui: 'notification-error',
-            title: headerText,
-            msg: errormsgs,
-            icon: Ext.MessageBox.ERROR
-        })
-    },*/
+    /*    handleFailureRequest: function (response, headerText, retryAction) {
+     Ext.EventManager.stopPropagation('requestexception');
+     var errormsgs = response.responseText;
+     Ext.widget('messagebox', {
+     buttons: [
+     {
+     text: 'Retry',
+     action: retryAction,
+     ui: 'remove'
+     },
+     {
+     text: 'Cancel',
+     action: 'cancel',
+     ui: 'link',
+     href: '#/administration/loadprofiletypes/',
+     handler: function (button, event) {
+     this.up('messagebox').hide();
+     }
+     }
+     ]
+     }).show({
+     ui: 'notification-error',
+     title: headerText,
+     msg: errormsgs,
+     icon: Ext.MessageBox.ERROR
+     })
+     },*/
 
 
     addMeasurementTypes: function () {
@@ -445,19 +437,11 @@ Ext.define('Mdc.controller.setup.LoadProfileTypes', {
     loadGridItemDetail: function (grid, record) {
         var form = Ext.ComponentQuery.query('loadProfileTypeSetup loadProfileTypePreview form')[0],
             previewPanel = this.getLoadTypePreview(),
-            recordData = record.getData(),
-            preloader = Ext.create('Ext.LoadMask', {
-                msg: "Loading...",
-                target: form
-            });
-        if (this.displayedItemId != recordData.id) {
-            grid.clearHighlight();
-            preloader.show();
-        }
+            recordData = record.getData();
+
         this.displayedItemId = recordData.id;
         previewPanel.setTitle(recordData.name);
         form.loadRecord(record);
-        preloader.destroy();
     },
 
     showLoadProfileTypes: function () {
